DT/FastAPI/requirements.txt
fastapi[standard]==0.95.0
uvicorn[standard]==0.21.1
numpy==1.23.5  # TensorFlow와 호환되는 버전
pandas==2.0.3
sqlalchemy==1.4.41  # databases와 호환되는 버전
tensorflow==2.12.0
pillow==9.4.0
requests==2.31.0
pydantic==1.10.8
databases==0.6.2
python-dotenv==1.0.0
motor==3.1.1
alembic==1.10.4  # SQLAlchemy와 호환되는 버전
apscheduler 